--- src/app/models/item.model.ts
@@ -1,7 +1,0 @@
-export interface Item{
-    id?: number;
-    name: string;
-    categoria: number;
-    desc: string;
-    stock: number;
-}
