+++ src/app/models/item.models.ts
@@ -1,0 +1,7 @@
+export interface Item{
+  id?: number;
+  name: string;
+  categoria: number;
+  desc: string;
+  stock: number;
+}
